[Silver II] Title: 주식, Time: 572 ms, Memory: 8292 KB -BaekjoonHub

+++ C++17/백준/Silver/11501. 주식/주식.cc
@@ -8,7 +8,7 @@ using namespace std;
 //11501
 
 int T, N, sellPrice;
-long long maxReturn;
+uint64_t maxReturn;
 vector<int> stock;
 
 int main(void){
